requirements.txt
python-dotenv
rich
gitpython
requests
langchain-gemini